requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
requests
pillow
